Extract database-related method
